# docs/swagger.yaml
definitions:
...
  requestmodels.CategoryOfferReq:
    properties:
      category_discount:
        maximum: 99
        minimum: 1
        type: integer
      category_id:
        type: string
      endDate:
        type: string
      restaurantID:
        type: string
      title:
        maxLength: 100
        type: string
      validityindays:
        minimum: 0
        type: integer
    required:
    - category_discount
    - category_id
    - title
    - validityindays
    type: object
  requestmodels.CategoryOfferStatus:
    properties:
      id:
        type: string
      status:
        maxLength: 20
        type: string
    required:
    - id
    - status
    type: object
  requestmodels.CategoryReq:
    properties:
      category_name:
        type: string
      categoryId:
        type: string
    required:
    - category_name
    type: object
...
  requestmodels.EditCategoryOffer:
    properties:
      category_discount:
        maximum: 99
        minimum: 1
        type: integer
      categoryOfferID:
        type: string
      endDate:
        type: string
      restaurantID:
        type: string
      title:
        type: string
      validityindays:
        type: integer
    required:
    - category_discount
    - title
    type: object
...
  responsemodels.CategoryRes:
    properties:
      category_id:
        type: string
      category_name:
        type: string
      category_status:
        type: string
    type: object
  responsemodels.Response:
    properties:
      after exicution: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
  description: Online Veg Food Delivery.
  title: Go + Gin Veg*n.
  version: "1.0"
paths:
...
  /admin/category:
    get:
      consumes:
      - application/json
      description: Retrieves all available categories.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responsemodels.CategoryRes'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - AdminRefTokenAuth: []
      summary: FetchAllCategory
      tags:
      - Category
    post:
      consumes:
      - application/json
      description: Creates a new category.
      parameters:
      - description: New category data.
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/requestmodels.CategoryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.CategoryRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - AdminRefTokenAuth: []
      summary: NewCategory
      tags:
      - Category
  /admin/category/{categoryid}:
    delete:
      consumes:
      - application/json
      description: Deletes a category.
      parameters:
      - description: The ID of the category to delete.
        in: path
        name: categoryid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - AdminRefTokenAuth: []
      summary: DeleteCategory
      tags:
      - Category
    patch:
      consumes:
      - application/json
      description: Updates an existing category.
      parameters:
      - description: The ID of the category to update.
        in: path
        name: categoryid
        required: true
        type: string
      - description: Updated category data.
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/requestmodels.CategoryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.CategoryRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - AdminRefTokenAuth: []
      summary: UpdateCategory
      tags:
      - Category
...
  /restaurant/category:
    get:
      consumes:
      - application/json
      description: Fetch all active categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - RestaurantAuthTokenAuth: []
      - RestaurantRefTokenAuth: []
      summary: Fetch active categories
      tags:
      - RestaurantCategoryManagement
  /restaurant/category/offer:
    get:
      consumes:
      - application/json
      description: Fetch all category offers for the restaurant specified in the request
        context
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - RestaurantAuthTokenAuth: []
      - RestaurantRefTokenAuth: []
      summary: Fetch all category offers
      tags:
      - RestaurantCategoryOfferManagement
    post:
      consumes:
      - application/json
      description: Create a new category offer for the restaurant specified in the
        request context
      parameters:
      - description: Category offer information
        in: body
        name: categoryOfferData
        required: true
        schema:
          $ref: '#/definitions/requestmodels.CategoryOfferReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - RestaurantAuthTokenAuth: []
      - RestaurantRefTokenAuth: []
      summary: Create a new category offer
      tags:
      - RestaurantCategoryOfferManagement
  /restaurant/category/offer/:
    patch:
      consumes:
      - application/json
      description: Change the status of a category offer for the restaurant specified
        in the request context
      parameters:
      - description: Category offer status information
        in: body
        name: categoryOfferStatusData
        required: true
        schema:
          $ref: '#/definitions/requestmodels.CategoryOfferStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - RestaurantAuthTokenAuth: []
      - RestaurantRefTokenAuth: []
      summary: Change the status of a category offer
      tags:
      - RestaurantCategoryOfferManagement
  /restaurant/category/offer/{categoryofferid}:
    patch:
      consumes:
      - application/json
      description: Edit a category offer for the restaurant specified in the request
        context
      parameters:
      - description: Category Offer ID to edit.
        in: path
        name: categoryofferid
        required: true
        type: string
      - description: Updated category offer information
        in: body
        name: categoryOfferData
        required: true
        schema:
          $ref: '#/definitions/requestmodels.EditCategoryOffer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - RestaurantAuthTokenAuth: []
      - RestaurantRefTokenAuth: []
      summary: Edit a category offer
      tags:
      - RestaurantCategoryOfferManagement
...
securityDefinitions:
  AdminRefTokenAuth:
    in: header
    name: refreshtoken
    type: apiKey
  OtpTempTokenAuth:
    in: header
    name: Authorizations
    type: apiKey
  RestaurantAuthTokenAuth:
    in: header
    name: accesstoken
    type: apiKey
  RestaurantRefTokenAuth:
    in: header
    name: refreshtoken
    type: apiKey
  UserAuthTokenAuth:
    in: header
    name: accesstoken
    type: apiKey
  UserRefTokenAuth:
    in: header
    name: refreshtoken
    type: apiKey
swagger: "2.0"